Extract file_helper.rs
